Add split images for nausea and test for symptom list
