+++ src/Patient/ReportingFlows/ReportSymptoms.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { observer } from 'mobx-react';
-import SymptomsList from './SymptomsList.js'
+import SymptomsList from './SymptomList/index.js'
 import SimpleButton from '../../Basics/SimpleButton'
 import useStores from '../../Basics/UseStores.js';
 import { Box, makeStyles } from '@material-ui/core';
